Latest smart-contract version

+++ src/main/java/com/dataart/maltahackaton/domain/Lottery.java
@@ -5,7 +5,12 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
@@ -42,7 +47,7 @@ public class Lottery {
 
     private BigDecimal ticketPrice;
 
-    private BigDecimal prizePoolRate;
+    private Long prizePoolRate;
 
     private byte[] lotteryPicture;
